important fix for generating basin areas from areas
if the geometry in areas is invalid it gets automatically corrected. This in turn can result in geometrycollections. Only the polygon or multipolygon in those collections will be used. In this filter step the identification of a geometrycollection was wrong resulting in incorrect basin areas from areas. has now been fixed by correctly identifying geometrycollection in filtering step